reorganize tests into folders
