+++ pytest.ini
@@ -1,12 +1,13 @@
 [pytest]
 testpaths = tests
-addopts =
-    --cov=fairy.cli
-    --cov=fairy.rulepack
-    --cov=fairy.validation.rulepack_runner
-    --cov-branch
-    --cov-report=term-missing
-    --cov-fail-under=35
+# Coverage options (require pytest-cov to be installed)
+# Uncomment these lines after running: pip install pytest-cov
+# addopts =
+#     --cov=fairy.cli
+#     --cov=fairy.rulepack
+#     --cov=fairy.validation.rulepack_runner
+#     --cov-branch
+#     --cov-report=term-missing
+#     --cov-fail-under=35
 filterwarnings =
     ignore::DeprecationWarning
-    ignore:Module .* was never imported:coverage.exceptions.CoverageWarning
