+++ tests/core/test_models/__init__.py
@@ -1,0 +1,1 @@
+# Core models tests
