+++ tests/core/test_models/test_manifest_role_inference.py
