+++ tests/core/test_services/__init__.py
@@ -1,0 +1,1 @@
+# Core services tests
