+++ tests/core/test_services/test_preflight_profiles_registry.py
