+++ tests/integration/test_cli_multi_input.py
