+++ tests/schema/test_schema_validation_edge_cases.py
@@ -18,7 +18,7 @@ from fairy.core.services.validator import run_rulepack
 
 def _get_schema_path() -> Path:
     """Get path to preflight_report_v1.schema.json relative to repo root."""
-    repo_root = Path(__file__).resolve().parent.parent
+    repo_root = Path(__file__).resolve().parents[2]
     schema_path = repo_root / "schemas" / "preflight_report_v1.schema.json"
     return schema_path
