+++ tests/validation/test_regex_rule_type.py
